adc tested

+++ Application/adc_read.h
@@ -3,5 +3,5 @@
 #include <stdint.h>
 #define ADC_READ_TIMEOUT_COUNT    200
 #define ADC_READ_TIMEOUT_RET_VALUE  0xFFFF
-uint16_t adc_read(int channel);
+unsigned int  adc_read(int channel);
 #endif
